requirements.txt
opencv-python
git+https://github.com/openai/CLIP.git
open_clip_torch
sentence_transformers
moviepy
pillow
tqdm